requirements.txt
pillow
tkinterdnd2
ttkbootstrap
